Split bq-jdbc-instrumented into bq-jdbc-tomcat-instrumented and bq-jdbc-hikaricp-instrumented #54
