+++ bootique-jdbc-hikaricp-instrumented/src/main/resources/META-INF/services/io.bootique.BQModuleProvider
@@ -1,0 +1,1 @@
+io.bootique.jdbc.hikaricp.instrumented.HikariCPInstrumentedModuleProvider
