--- bootique-jdbc-instrumented/src/main/resources/META-INF/services/io.bootique.BQModuleProvider
@@ -1,1 +1,0 @@
-io.bootique.jdbc.instrumented.InstrumentedJdbcModuleProvider
