--- bootique-jdbc-instrumented/src/test/java/io/bootique/jdbc/instrumented/InstrumentedJdbcModuleProviderTest.java
@@ -1,19 +1,0 @@
-package io.bootique.jdbc.instrumented;
-
-import io.bootique.test.junit.BQModuleProviderChecker;
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-
-public class InstrumentedJdbcModuleProviderTest {
-
-	@Test
-	public void testPresentInJar() {
-		BQModuleProviderChecker.testPresentInJar(InstrumentedJdbcModuleProvider.class);
-	}
-
-	@Test
-	public void testConfigPrefix() {
-		assertEquals("jdbc", new InstrumentedJdbcModule().getConfixPrefix());
-	}
-}
