+++ bootique-jdbc-test/pom.xml
@@ -18,7 +18,7 @@
 		<dependencies>
 			<dependency>
 				<groupId>io.bootique.jdbc</groupId>
-				<artifactId>bootique-jdbc-instrumented</artifactId>
+				<artifactId>bootique-jdbc-tomcat-instrumented</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 			<dependency>
@@ -57,7 +57,7 @@
 		<!-- Compile dependencies -->
 		<dependency>
 			<groupId>io.bootique.jdbc</groupId>
-			<artifactId>bootique-jdbc-instrumented</artifactId>
+			<artifactId>bootique-jdbc-tomcat-instrumented</artifactId>
 			<scope>compile</scope>
 		</dependency>
 		<dependency>
